Project files reorganization (+ fixed reloading)
