+++ frontend/static/hbs/home.hbs
@@ -9,7 +9,10 @@
       href="https://fonts.googleapis.com/css2?family=Suez+One&display=swap"
       rel="stylesheet"
     />
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css">
+    <link
+      rel="stylesheet"
+      href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.3.0/css/all.min.css"
+    />
     <link
       href="https://fonts.googleapis.com/css2?family=Poppins:wght@100;200;300;400;500;600;700;800&display=swap"
       rel="stylesheet"
@@ -54,7 +57,7 @@
           <img class="icon" src="/static/media/profile.svg" />
           <div class="text">Profile</div>
         </a>
-        <a class="nav-item logout" href="/logout">
+        <a class="nav-item logout" href="/logout" action="/logout" method="GET">
           <img class="icon" src="/static/media/logout.svg" />
           <div class="text">Logout</div>
         </a>
