+++ frontend/static/hbs/register.hbs
@@ -34,10 +34,23 @@
     <div class="container mt-4">
       <div class="card">
         <div class="card-header">
-          Login
+          Registration
         </div>
         <div class="card-body">
-          <form action="/auth/login" method="POST">
+          <form action="/register" method="POST">
+            <div class="mb-3">
+              <label for="name" class="form-label">Name</label>
+              <input type="name" class="form-control" id="name" name="name" />
+            </div>
+            <div class="mb-3">
+              <label for="surname" class="form-label">Surname</label>
+              <input
+                type="surname"
+                class="form-control"
+                id="surname"
+                name="surname"
+              />
+            </div>
             <div class="mb-3">
               <label for="email" class="form-label">Email</label>
               <input
@@ -56,20 +69,37 @@
                 name="password"
               />
             </div>
-            <button type="Submit" class="btn btn-primary">Login</button>
+            <div class="mb-3">
+              <label for="passwordConfirm" class="form-label">Confirm password</label>
+              <input
+                type="password"
+                class="form-control"
+                id="passwordConfirm"
+                name="passwordConfirm"
+              />
+            </div>
+            <label for="role" class="mb-3">Gender</label>
+            <select id="role" name="role" class="form-control mb-3">
+              <option value="">Select Role</option>
+              <option value="project-manager">Project manager</option>
+              <option value="developer">developer</option>
+            </select>
+
+            <button type="submit" class="btn btn-primary">Submit</button>
             <div class="text-center mt-3">
-              <span>Have an invitation code? </span>
-              <a href="/register" class="link-primary">Register</a>
+              <span>Already have an account? </span>
+              <a href="/login" class="link-primary">Login</a>
             </div>
             {{#if message}}
-              <h4 class="alert mt-4">{{message}}</h4>
+              <div class="alert mt-4">{{message}}</div>
             {{/if}}
-
           </form>
         </div>
+
       </div>
     </div>
 
+    <script type="module" src="/static/js/index.js"></script>
     <script
       src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha1/dist/js/bootstrap.bundle.min.js"
       integrity="sha384-w76AqPfDkMBDXo30jS1Sgez6pr3x5MlQ1ZAGC+nuZB+EYdgRZgiwxhTBTkF7CXvN"
