+++ server.js
@@ -1,32 +1,28 @@
+const PORT = 3000;
 const express = require("express");
 const path = require("path");
 const mysql = require("mysql");
 const dotenv = require("dotenv");
-const cookieParser = require("cookie-parser");
-const authController = require("./controllers/auth");
 dotenv.config({ path: "./.env" });
-
+const cookieParser = require("cookie-parser");
+const hbs = require("hbs");
 const app = express();
-
 const db = mysql.createConnection({
   host: process.env.HOST,
   user: process.env.DATABASE_USER,
   password: process.env.DATABASE_PASSWORD,
   database: process.env.DATABASE,
 });
+const root = __dirname; //__dirname visada yra ta direktorija kur yra jis pats iskviestas siuo atveju __dirname yra ten kur server.js failas
+const backendDir = path.join(root, "backend");
+const frontendDir = path.join(root, "frontend");
+const staticDir = path.join(frontendDir, "static");
+const routesDir = path.join(backendDir, "routes");
 
-//const publicDirectory = path.join(__dirname, "./frontend/static");
+//Define routes
 
 app.set("view engine", "hbs");
-
-app.set("views", __dirname + "/views");
-
-// pasiemam stilius ir views
-app.use(
-  "/static",
-  express.static(path.resolve(__dirname, "frontend", "static"))
-);
-app.use("/media", express.static(path.resolve(__dirname, "frontend", "media")));
+app.set("views", path.join(staticDir, "hbs"));
 
 //Parse URL-encoded bodies (as sent by HTML forms)
 app.use(express.urlencoded({ extend: true }));
@@ -45,16 +41,18 @@ db.connect((error) => {
     console.log("MYSQL Connected...");
   }
 });
+module.exports = {
+  db: db,
+};
 
-app.get("/", authController.home);
-app.get("/logout", authController.logout);
-app.get("/login", authController.renderLogin);
+app.use("/login", require(path.join(routesDir, "login")));
+app.use("/register", require(path.join(routesDir, "register")));
+app.use("/", require(path.join(routesDir, "dashboard")));
 
-//Define routes
-app.use("/", require("./routes/pages"));
-app.use("/auth", require("./routes/auth"));
+app.use("/media", express.static(path.join(staticDir, "media")));
+app.use("/static", express.static(staticDir));
 
 //nusistatom porta
-app.listen(process.env.PORT || 3000, () =>
+app.listen(process.env.PORT || PORT, () =>
   console.log("Dreamwork has started, the server is running on 3000 port")
 );
